requirements.txt
celery-5.3.1
django-4.2.2
djangorestframework-3.14.0
psycopg2-binary--2.9.6
djangorestframework_simplejwt-4.4.0
django-redis-5.3.0
django-celery-beat-2.5.0
django-celery-results-2.5.1